+++ swissgeodownloader/__init__.py
@@ -18,15 +18,17 @@
  *                                                                         *
  ***************************************************************************/
 """
+
 import os
 
-__version__ = '2.2.0'
+__version__ = "2.2.0"
 
 DEBUG = False
 PLUGIN_DIR = os.path.dirname(__file__)
-_AVAILABLE_LOCALES = ['de', 'en', 'fr', 'it']
+_AVAILABLE_LOCALES = ["de", "en", "fr", "it"]
 
 
 def classFactory(iface):
     from .swissgeodownloader import SwissGeoDownloader
+
     return SwissGeoDownloader(iface)