+++ swissgeodownloader/scripts/datageoadmin_catalog_analysis.py
@@ -9,21 +9,22 @@ from swissgeodownloader.utils.metadata_handler import saveToFile
 
 def refreshMetadata():
     api = ApiDataGeoAdmin()
-    task = ApiCallerTask(api, None, '')
+    task = ApiCallerTask(api, None, "")
     api.refreshAllMetadata(task)
 
 
 def analyseFullCatalog():
-    api = ApiDataGeoAdmin('de')
-    task = ApiCallerTask(api, None, '')
+    api = ApiDataGeoAdmin("de")
+    task = ApiCallerTask(api, None, "")
     items = api.catalogPropertiesCrawler(task)
     save(items)
 
 
 def save(items):
     if items:
-        savePath = os.path.join(os.path.dirname(__file__),
-                                'datageoadmin_catalog_analysis.json')
+        savePath = os.path.join(
+            os.path.dirname(__file__), "datageoadmin_catalog_analysis.json"
+        )
         saveToFile(items, savePath)
 
 
@@ -31,16 +32,16 @@ def onError(error):
     print(error)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     # Get arguments from command line
     import sys
-    
+
     action = sys.argv[1] if len(sys.argv) > 1 else "analyseFullCatalog"
-    
+
     QGIS_APP = QgsApplication([], False)
     QGIS_APP.initQgis()
-    
-    if action == 'refreshMetadata':
+
+    if action == "refreshMetadata":
         refreshMetadata()
-    elif action == 'analyseFullCatalog':
+    elif action == "analyseFullCatalog":
         analyseFullCatalog()