+++ swissgeodownloader/ui/collection_list_table.py
@@ -18,15 +18,22 @@
  *                                                                         *
  ***************************************************************************/
 """
+
 from qgis.PyQt.QtCore import (
-    QCoreApplication, QObject, QSortFilterProxyModel,
-    Qt, pyqtSignal
+    QCoreApplication,
+    QObject,
+    QSortFilterProxyModel,
+    Qt,
+    pyqtSignal,
 )
 from qgis.PyQt.QtGui import QStandardItem, QStandardItemModel
 from qgis.PyQt.QtWidgets import (
-    QAbstractItemView, QAbstractScrollArea,
-    QHeaderView, QLineEdit, QSizePolicy,
-    QTableView
+    QAbstractItemView,
+    QAbstractScrollArea,
+    QHeaderView,
+    QLineEdit,
+    QSizePolicy,
+    QTableView,
 )
 
 from swissgeodownloader.api.response_objects import SgdStacCollection
@@ -34,15 +41,16 @@ from swissgeodownloader.api.response_objects import SgdStacCollection
 
 class CollectionListTable(QObject):
     sig_selectionChanged = pyqtSignal(str)
-    
+
     def __init__(self, parent, layout):
         super().__init__()
         self.parent = parent
         self.currentSelection = None
-        
+
         self.tbl = QTableView(self.parent)
-        sizePolicy = QSizePolicy(QSizePolicy.Policy.Expanding,
-                                 QSizePolicy.Policy.MinimumExpanding)
+        sizePolicy = QSizePolicy(
+            QSizePolicy.Policy.Expanding, QSizePolicy.Policy.MinimumExpanding
+        )
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(1)
         sizePolicy.setHeightForWidth(self.tbl.sizePolicy().hasHeightForWidth())
@@ -57,9 +65,11 @@ class CollectionListTable(QObject):
         self.tbl.setSizePolicy(sizePolicy)
         self.tbl.setMinimumHeight(90)
         self.tbl.setMaximumHeight(250)
-        self.tbl.setSizeAdjustPolicy(QAbstractScrollArea.SizeAdjustPolicy.AdjustToContents)
+        self.tbl.setSizeAdjustPolicy(
+            QAbstractScrollArea.SizeAdjustPolicy.AdjustToContents
+        )
         self.tbl.setAutoScroll(True)
-        
+
         self.tbl.setSelectionMode(QAbstractItemView.SelectionMode.SingleSelection)
         self.tbl.setSelectionBehavior(QAbstractItemView.SelectionBehavior.SelectRows)
         self.tbl.setObjectName("DatasetListTable")
@@ -70,17 +80,17 @@ class CollectionListTable(QObject):
 
         self.searchbar = QLineEdit()
         self.searchbar.setClearButtonEnabled(True)
-        self.searchbar.setPlaceholderText(self.tr('Search'))
-        
+        self.searchbar.setPlaceholderText(self.tr("Search"))
+
         layout.addWidget(self.searchbar)
         layout.addWidget(self.tbl)
 
         self.searchbar.textChanged.connect(self.onSearch)
         self.tbl.clicked.connect(self.onClick)
-    
+
     def fill(self, data: list[SgdStacCollection]):
         self.model.clear()
-        
+
         # Insert data into cells
         for i, coll in enumerate(data):
             item0 = QStandardItem(coll.id())
@@ -95,13 +105,14 @@ class CollectionListTable(QObject):
             self.model.setData(self.model.index(i, 0), coll.id())
             self.model.setData(self.model.index(i, 1), coll.title())
             self.model.setData(self.model.index(i, 2), coll.searchText())
-            
-        self.model.setHorizontalHeaderLabels([self.tr('ID'), self.tr('Title'),
-                                              self.tr('Search text')])
+
+        self.model.setHorizontalHeaderLabels(
+            [self.tr("ID"), self.tr("Title"), self.tr("Search text")]
+        )
         self.tbl.setColumnWidth(0, 130)
         # Search text is not visible
         self.tbl.hideColumn(2)
-    
+
     def onSearch(self, search):
         self.proxy_model.setFilterFixedString(search.lower())
         # Remove selection if the selected item is not visible any more
@@ -116,7 +127,7 @@ class CollectionListTable(QObject):
         else:
             self.currentSelection = dsId
             self.sig_selectionChanged.emit(dsId)
-    
+
     def searchAndSelectByID(self, collectionId):
         self.resetSearch()
         self.tbl.clearSelection()
@@ -124,15 +135,14 @@ class CollectionListTable(QObject):
         self.tbl.selectRow(0)
         self.currentSelection = collectionId
         self.sig_selectionChanged.emit(collectionId)
-    
+
     def unselect(self):
         self.tbl.clearSelection()
         self.currentSelection = None
         self.sig_selectionChanged.emit(None)
-    
+
     def resetSearch(self):
         self.searchbar.clear()
 
     def tr(self, message, **kwargs):
         return QCoreApplication.translate(type(self).__name__, message)
-