+++ swissgeodownloader/utils/qgis_layer_creator_task.py
@@ -18,6 +18,7 @@
  *                                                                         *
  ***************************************************************************/
 """
+
 import os
 
 from osgeo import gdal
@@ -27,7 +28,7 @@ from qgis.core import (
     QgsRasterLayer,
     QgsTask,
     QgsVectorLayer,
-    Qgis
+    Qgis,
 )
 
 from swissgeodownloader.api.response_objects import SgdAsset
@@ -38,21 +39,21 @@ def createQgisLayersInTask(fileList: list[SgdAsset], vrtOutputPath, callback):
     # Create layer from files (streamed and downloaded) so they can be
     # added to qgis
     task = QgisLayerCreatorTask(
-            translate('SGD', 'Adding files to QGIS...'),
-            fileList, vrtOutputPath)
+        translate("SGD", "Adding files to QGIS..."), fileList, vrtOutputPath
+    )
     task.taskCompleted.connect(
-            lambda: callback(task.layerList, task.alreadyAdded,
-                             task.exception))
+        lambda: callback(task.layerList, task.alreadyAdded, task.exception)
+    )
     task.taskTerminated.connect(lambda: callback([], 0, task.exception))
     QgsApplication.taskManager().addTask(task)
-    
-    
+
+
 class QgisLayerCreatorTask(QgsTask):
-    """ QGIS can freeze when a lot of layers have to be created in the main
-     thread. Instead, layers are created in this separate QTask and moved to
-     the main thread. After the task has finished, they are added to the map
-     in the main thread."""
-    
+    """QGIS can freeze when a lot of layers have to be created in the main
+    thread. Instead, layers are created in this separate QTask and moved to
+    the main thread. After the task has finished, they are added to the map
+    in the main thread."""
+
     def __init__(self, description, fileList: list[SgdAsset], vrtOutputPath):
         super().__init__(description, QgsTask.Flag.CanCancel)
         self.fileList = fileList
@@ -60,46 +61,45 @@ class QgisLayerCreatorTask(QgsTask):
         self.alreadyAdded: int = 0
         self.exception = None
         self.vrtOutputPath = vrtOutputPath
-    
+
     def run(self):
-        
+
         if not self.fileList or len(self.fileList) == 0:
             return True
-        
+
         qgsProject = QgsProject.instance()
-        already_added = [lyr.source() for lyr in
-                         qgsProject.mapLayers().values()]
-        
+        already_added = [lyr.source() for lyr in qgsProject.mapLayers().values()]
+
         progressStep = 100 / len(self.fileList)
-        
+
         if self.vrtOutputPath:
             return self.combineTiles()
-        
+
         for i, file in enumerate(self.fileList):
             if self.isCanceled():
                 return False
-            
+
             self.setProgress(i * progressStep)
-            
+
             # Adding the file to QGIS if it's (1) a streamed file or (2) is
             #  present in the file system and (3) is not a .zip
             if file.isStreamable or os.path.exists(file.path):
-                if '.zip' in file.id:
+                if ".zip" in file.id:
                     # Can't add zip files to QGIS
                     continue
                 if file.path in already_added:
                     self.alreadyAdded += 1
                     continue
-                
+
                 # Try to create a raster layer first, if that fails, create a
                 #  vector layer
                 success = self.createRasterLayer(file.path, file.id)
                 if success:
                     continue
-                
+
                 self.createVectorLayer(file.path, file.id)
         return True
-    
+
     def combineTiles(self):
         try:
             pathList = [file.path for file in self.fileList]
@@ -109,7 +109,7 @@ class QgisLayerCreatorTask(QgsTask):
             return False
         layerName = os.path.splitext(os.path.basename(self.vrtOutputPath))[0]
         return self.createRasterLayer(self.vrtOutputPath, layerName)
-    
+
     def createRasterLayer(self, filepath, filename):
         try:
             rasterLyr = QgsRasterLayer(filepath, filename)
@@ -119,9 +119,13 @@ class QgisLayerCreatorTask(QgsTask):
             else:
                 del rasterLyr
         except Exception as e:
-            self.exception = f'{filename}: {e}' if self.exception is None else self.exception + f'\n{filename}: {e}'
+            self.exception = (
+                f"{filename}: {e}"
+                if self.exception is None
+                else self.exception + f"\n{filename}: {e}"
+            )
         return False
-    
+
     def createVectorLayer(self, filepath, filename):
         try:
             vectorLyr = QgsVectorLayer(filepath, filename, "ogr")
@@ -131,15 +135,19 @@ class QgisLayerCreatorTask(QgsTask):
             else:
                 del vectorLyr
         except Exception as e:
-            self.exception = f'{filename}: {e}' if self.exception is None else self.exception + f'\n{filename}: {e}'
+            self.exception = (
+                f"{filename}: {e}"
+                if self.exception is None
+                else self.exception + f"\n{filename}: {e}"
+            )
         return False
-    
+
     def finished(self, result):
         self.setProgress(100)
         if not result:
             if self.isCanceled():
-                self.exception = self.tr('Aborted by user')
+                self.exception = self.tr("Aborted by user")
             elif self.exception is None:
-                self.exception = self.tr('An unknown error occurred')
-        for e in str(self.exception).split('\n'):
+                self.exception = self.tr("An unknown error occurred")
+        for e in str(self.exception).split("\n"):
             log(e, Qgis.MessageLevel.Warning)